+++ src/sdcard.cpp
@@ -12,16 +12,10 @@ SPIClass spi = SPIClass(VSPI);
 
 String addCRLF(const String& str) {
   String result = str;
-//   Serial.print(str);
-//   Serial.print(result.substring(result.length() - 2));
-//   if (result.substring(result.length() - 1) == "\n") {
-//     // result.replace("\n", "\r\n");
-//   } else 
   if (result.length() < 2 || result.substring(result.length() - 2) != "\r\n") {
     if (result.substring(result.length() - 1) == "\n") {
         result.replace("\n", "\r\n");
     } else {
-        // result += "\r\n";
         result.concat("\r\n");
     }
   }
@@ -33,9 +27,6 @@ bool searchForFile(fs::FS &fs, const char* fileName) {
     File root = fs.open("/");
     File file = root.openNextFile();
     while (file) {
-        // if (file.isDirectory()) {
-        //     continue;   // skip directories
-        // }
         if (strcmp(file.name(), fileName) == 0) {
             Serial.print("Found file: ");
             Serial.println(file.name());
@@ -156,37 +147,6 @@ bool deleteFile(fs::FS &fs, const char * path){
         return(fs.remove(path));
     }
 }
-
-// bool writeGPXpoint(const char * fileName, int wptNum, double lat, double lon){
-//     String waypoint;
-
-//     waypoint = "<trkpt lat=\"";
-//     waypoint += String(lat,6);
-//     waypoint += "\" lon=\"";
-//     waypoint += String(lon,6);
-//     waypoint += "\">\n  <name>";
-//     waypoint += wptNum;
-//     waypoint += "</name>\n</trkpt>\n";
-
-//     return (appendFile(SD, fileName, waypoint.c_str(), false));
-// }
-
-// bool createGPXfile(const char * fileName){
-//     String header;
-
-//     header = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<gpx version=\"1.1\" creator=\"NMEATrax\">\n";
-//     header += "<trk>\n<name>";
-//     header += fileName;
-//     header += "</name>\n<trkseg>\n";
-
-//     return (appendFile(SD, fileName, header.c_str(), false));
-// }
-
-// bool endGPXfile(const char * fileName){
-//     String footer;
-//     footer = "</trkseg>\n</trk>\n</gpx>\n";
-//     return (appendFile(SD, fileName, footer.c_str(), false));
-// }
 
 bool sdSetup(){
     const uint8_t SCK = 18;