+++ changelog.md
@@ -9,10 +9,16 @@ and this project tries to adhere to [Semantic Versioning](http://semver.org/).
  
 ### Coming Soon
 - Fix WiFi settings not saving correctly
-- Add GPS nav msgs
-- TimeString to epoch
 
-### In Progress
+
+## [9.0.0] - 2024-08-10
+
+### Added
+- GPS Navigation messages
+
+### Changed
+- TimeString is now sent as unixtime.
+- Units in metric always. Client to handle conversion.
 
 
 ## [8.0.0] - 2024-07-25